@@ -1,6 +1,6 @@
 {
   "name": "ionic-cache",
-  "version": "3.1.3",
+  "version": "4.0.1",
   "description": "Ionic cache service - cache request, data, promises etc.",
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
@@ -13,7 +13,8 @@
     "test:watch": "karma start karma.conf.js",
     "build": "rimraf dist && ngc",
     "lint": "tslint \"src/**/*.ts\"",
-    "prepublishOnly": "npm run lint && npm run build"
+    "prepublishOnly": "npm run lint && npm run build",
+    "prepare": "npm run build"
   },
   "repository": {
     "type": "git",
@@ -32,14 +33,14 @@
   },
   "homepage": "https://github.com/Nodonisko/ionic-cache#readme",
   "devDependencies": {
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
     "@ionic/storage": "^2.1.3",
     "@types/jasmine": "^2.5.47",
     "@types/node": "^7.0.18",
@@ -54,16 +55,13 @@
     "karma-typescript-angular2-transform": "^1.0.2",
     "karma-typescript-es6-transform": "^1.0.4",
     "rimraf": "^2.6.2",
-    "rxjs": "^5.5.0",
+    "rxjs": "6.2.2",
     "tslint": "^3.15.1",
     "tslint-ionic-rules": "^0.0.8",
-    "typescript": "~2.6.0",
+    "typescript": "~2.7.2",
     "zone.js": "^0.8.11"
   },
   "peerDependencies": {
     "@ionic/storage": "^2.1.3"
-  },
-  "browser": {
-    "@angular/common/http": false
   }
 }
